+++ src/models/transaction/imported.py
@@ -1,5 +1,4 @@
 from dataclasses import dataclass
-from decimal import Decimal
 from typing import Tuple
 
 
@@ -9,14 +8,14 @@ class ImportedTransaction:
 
     account_number: str
 
-    opening_balance: Decimal
+    opening_balance: float
     closing_balance: Tuple[str, str, str]
 
     date: str
     booking_date: str
 
     currency: str
-    amount: Decimal
+    amount: float
 
     transaction_type_number: str
     transaction_type_name: str
